QEMU orchestrator: Respect the `keep_environment` flag. (#1730)
* QEMU orchestrator: Respect the `keep_environment` flag.

If the `keep_environment` flag has been set, then don't
delete the VMs if the VM deployment fails.

+++ lisa/sut_orchestrator/qemu/platform.py
@@ -103,7 +103,13 @@ class QemuPlatform(Platform):
                 self._fill_nodes_metadata(environment, log, qemu_conn)
 
             except Exception as ex:
-                self._delete_nodes(environment, log, qemu_conn)
+                assert environment.platform
+                if (
+                    environment.platform.runbook.keep_environment
+                    == constants.ENVIRONMENT_KEEP_NO
+                ):
+                    self._delete_nodes(environment, log, qemu_conn)
+
                 raise ex
 
     # Pre-determine all the nodes' properties, including the name of all the resouces
